graphql-alt: Scope.root_version
## Description

Introduce the concept of a "root version" to the scope. This version is
used as a version upperbound by nested queries for child objects
(dynamic fields). It is set whenever an object is fetched at a specific
version, or with a version-bound.

This change also applies limits to the interplay between
versioon-bounded queries (which now set the root version), and owned
object queries (which need to be pinned to a particular checkpoint).

## Test plan

New E2E tests:

```
$ cargo nextest run -p sui-indexer-alt-e2e-tests
```